Visual adjustments.

+++ core/domain_settings/domain_settings.php
@@ -275,7 +275,7 @@ if (sizeof($_REQUEST) > 1) {
 	if (permission_exists('domain_setting_add')) {
 		echo 		"<a href='domain_setting_edit.php?domain_uuid=".check_str($_GET['id'])."' alt='".$text['button-add']."'>$v_link_label_add</a>";
 	}
-	if (permission_exists('domain_setting_delete')) {
+	if (permission_exists('domain_setting_delete') && $result_count > 0) {
 		echo "<a href='javascript:void(0);' onclick=\"if (confirm('".$text['confirm-delete']."')) { document.getElementById('action').value = 'delete'; document.getElementById('domain_frm').submit(); }\" alt='".$text['button-delete']."'>".$v_link_label_delete."</a>";
 	}
 	echo "		</td>\n";
